requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Mail==0.9.1
python-dotenv==0.17.1
SQLAlchemy==1.3.23


